requirements.txt
asgiref
certifi
charset-normalizer
dj-database-url
Django
django-cors-headers
django-environ
djangorestframework
djangorestframework_simplejwt
factory-boy
idna
joblib
numpy
pandas
pillow
psycopg2
PyJWT
python-dateutil
python-decouple
pytz
requests
scikit-learn
scipy
setuptools
six
sqlparse
sslcommerz-lib
threadpoolctl
typed_ast
typing_extensions
tzdata
urllib3
wheel
gunicorn
whitenoise
